requirements.txt
streamlit
fpdf
python-docx
beautifulsoup4
requests
